+++ src/runmd/process.py
@@ -105,13 +105,14 @@ def show_code_block(name, lang, code, tag):
     except Exception as e:
         print(f"Error: Code block '{name}' failed with exception: {e}")
 
-def run_command(blocklist: list, block_name: str, config: dict, env_vars: dict) -> None:
+def run_command(blocklist: list, block_name: str, tag: str, config: dict, env_vars: dict) -> None:
     """
     Handle the 'run' command to execute code blocks.
 
     Args:
         code_blocks (list): List of code blocks extracted from Markdown.
         block_name (str): Name of the code block to run or 'all' to run all.
+        tag(str): Name of the tag of the code blocks to execute
         config (dict): Configuration dictionary.
 
     Returns:
@@ -120,7 +121,7 @@ def run_command(blocklist: list, block_name: str, config: dict, env_vars: dict) 
 
     block_count = 0
     for block in blocklist:
-        if block_name == "all" or block_name == block['name'] or '@'+block['tag'] == block_name:
+        if block_name == "all" or block_name == block['name'] or tag == block['tag']:
             if block['exec']:
                 run_code_block(block['name'], block['lang'], block['code'], block['tag'],config, env_vars)
                 block_count +=1
@@ -129,4 +130,7 @@ def run_command(blocklist: list, block_name: str, config: dict, env_vars: dict) 
                 block_count +=1
     
     if block_name != "all" and block_count == 0:
-        print(f"Error: Code block with name '{block_name}' not found.")
+        if tag is not None:
+            print(f"Error: Code block with tag '{tag}' not found.")
+        else:
+            print(f"Error: Code block with name '{block_name}' not found.")